Bug fixes and implementation testing with Windows

+++ src/pkgmngr/apparatus.py
@@ -5,8 +5,8 @@ from subprocess import check_output
 class Apparatus:
     SETTINGS = dict()
     #defines path to working dir of 'legoHDL' tool
-    PKGMNG_PATH = os.path.realpath(__file__)[:os.path.realpath(__file__).rfind('/')+1]
-    #path to registry and cache
+    PKGMNG_PATH = os.path.realpath(__file__)[:os.path.realpath(__file__).replace("\\","/").rfind('/')+1]
+    #path to registry and cachess
     HIDDEN = os.path.expanduser("~/.legohdl/")
 
     WORKSPACE = HIDDEN
@@ -14,35 +14,72 @@ class Apparatus:
     __active_workspace = None
 
     @classmethod
+    def initialize(cls):
+        os.makedirs(cls.HIDDEN, exist_ok=True)
+        os.makedirs(cls.HIDDEN+"workspaces/", exist_ok=True)
+        os.makedirs(cls.HIDDEN+"scripts/", exist_ok=True)
+        os.makedirs(cls.HIDDEN+"registry/", exist_ok=True)
+        os.makedirs(cls.HIDDEN+"template/", exist_ok=True)
+        #create bare settings.yml if DNE
+        if(not os.path.isfile(cls.HIDDEN+"settings.yml")):
+            settings_file = open(cls.HIDDEN+"settings.yml", 'w')
+            structure = """active-workspace:
+author:
+editor:
+label:
+  recursive: {}
+  shallow: {}
+market: {}
+script: {}
+workspace: {}            
+            """
+            settings_file.write(structure)
+            settings_file.close()
+
+    @classmethod
     def load(cls):
+        cls.initialize()
         log.basicConfig(format='%(levelname)s:\t%(message)s', level=log.DEBUG)
 
-        with open(cls.PKGMNG_PATH+"settings.yml", "r") as file:
+        with open(cls.HIDDEN+"settings.yml", "r") as file:
             cls.SETTINGS = yaml.load(file, Loader=yaml.FullLoader)
         
         #determine current workspace currently being used
         cls.__active_workspace = cls.SETTINGS['active-workspace']
 
-        if(cls.__active_workspace == None or cls.__active_workspace not in cls.SETTINGS['workspace'].keys()):
-            exit("ERROR- Active workspace not found!")
+        if(not cls.inWorkspace()):
+            log.warning("Active workspace not found!")
+            return
         
         if(cls.SETTINGS['workspace'][cls.__active_workspace]['local'] == None):
-            exit("ERROR- Please specify a local path! See \'legohdl help config\' for more details")
+            log.error("Please specify a local path! See \'legohdl help config\' for more details")
 
         cls.WORKSPACE = cls.HIDDEN+"workspaces/"+cls.SETTINGS['active-workspace']+"/"
         pass
 
     @classmethod
+    def inWorkspace(cls):
+        #determine current workspace currently being used
+        cls.__active_workspace = cls.SETTINGS['active-workspace']
+        if(cls.__active_workspace == None or cls.__active_workspace not in cls.SETTINGS['workspace'].keys()):
+            return False
+        else:
+            return True
+
+    @classmethod
     def initializeWorkspace(cls, name):
         workspace_dir = cls.HIDDEN+"workspaces/"+name+"/"
         os.makedirs(workspace_dir, exist_ok=True)
-        os.mkdir(workspace_dir+"lib")
-        os.mkdir(workspace_dir+"cache")
-        open(workspace_dir+"map.toml", 'w').write("[libraries]\n")
+        os.makedirs(workspace_dir+"lib", exist_ok=True)
+        os.makedirs(workspace_dir+"cache", exist_ok=True)
+        if(not os.path.isfile(workspace_dir+"map.toml")):
+            open(workspace_dir+"map.toml", 'w').write("[libraries]\n")
+        if(not cls.inWorkspace()):
+            cls.SETTINGS['active-workspace'] = name
     
     @classmethod
     def save(cls):
-        with open(cls.PKGMNG_PATH+"settings.yml", "w") as file:
+        with open(cls.HIDDEN+"settings.yml", "w") as file:
             yaml.dump(cls.SETTINGS, file)
         pass
     
@@ -54,8 +91,9 @@ class Apparatus:
     def getMarkets(cls):
         returnee = dict()
         #key: name, val: url
-        for name in cls.SETTINGS['workspace'][cls.__active_workspace]['market']:
-            returnee[name] = cls.SETTINGS['market'][name]
+        if(cls.inWorkspace()):
+            for name in cls.SETTINGS['workspace'][cls.__active_workspace]['market']:
+                returnee[name] = cls.SETTINGS['market'][name]
         return returnee
 
     @classmethod
